+++ packages/experimental/framework/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/framework
 
-This log was last generated on Wed, 19 Aug 2020 19:19:35 GMT and should not be manually modified.
+This log was last generated on Thu, 20 Aug 2020 00:14:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.10
+
+Thu, 20 Aug 2020 00:14:01 GMT
+
+### Patches
+
+- update code to use split out merge-props package (jasonmo@microsoft.com)
 
 ## 0.2.9